--- src/main/java/com/thinkmore/forum/dto/oauth/OauthPostDto.java
@@ -1,13 +1,0 @@
-package com.thinkmore.forum.dto.oauth;
-
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-import java.io.Serializable;
-
-@Data
-@NoArgsConstructor
-public class OauthPostDto implements Serializable {
-    private String oauthType;
-    private String openid;
-}
