+++ src/main/java/com/thinkmore/forum/service/UsersService.java
@@ -1,28 +1,18 @@
 package com.thinkmore.forum.service;
 
-import com.thinkmore.forum.dto.oauth.OauthPostDto;
 import com.thinkmore.forum.dto.users.*;
 import com.thinkmore.forum.entity.JwtUser;
 import com.thinkmore.forum.configuration.StaticConfig;
-import com.thinkmore.forum.entity.Oauth;
 import com.thinkmore.forum.entity.Roles;
 import com.thinkmore.forum.entity.Users;
 import com.thinkmore.forum.exception.InvalidOldPasswordException;
 import com.thinkmore.forum.exception.UserNotFoundException;
 import com.thinkmore.forum.mapper.UsersMapper;
-import com.thinkmore.forum.entity.rabbitmq.ResetPasswordEmailMessage;
-import com.thinkmore.forum.entity.rabbitmq.VerificationEmailMessage;
-import com.thinkmore.forum.repository.OauthRepository;
 import com.thinkmore.forum.repository.RolesRepository;
 import com.thinkmore.forum.repository.UsersRepository;
-import com.thinkmore.forum.configuration.RabbitMinioConfig;
-import com.thinkmore.forum.util.Util;
 
 import lombok.RequiredArgsConstructor;
-import org.springframework.amqp.rabbit.annotation.RabbitListener;
-import org.springframework.amqp.rabbit.core.RabbitTemplate;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
@@ -41,17 +31,10 @@ import java.util.stream.Collectors;
 @RequiredArgsConstructor
 public class UsersService implements UserDetailsService {
 
-    @Value("${domain.name}")
-    public String domainName;
-
-    @Autowired
-    RabbitTemplate rabbitTemplate;
-
     @Autowired
     PasswordEncoder passwordEncoder;
 
     private final UsersRepository usersRepository;
-    private final OauthRepository oauthRepository;
     private final UsersMapper usersMapper;
     private final RolesRepository rolesRepository;
 
@@ -78,43 +61,6 @@ public class UsersService implements UserDetailsService {
     }
 
     @Transactional
-    public Boolean thirdPartyLogin(String email, String username, OauthPostDto oauthPostDto) {
-        if (!uniqueEmail(email)) {
-            Users users = usersRepository.findByEmail(email).get();
-            users.setPassword(passwordEncoder.encode(oauthPostDto.getOpenid()));
-
-            Oauth oauth = new Oauth();
-
-            oauth.setUsers(users);
-            oauth.setOauthType(oauthPostDto.getOauthType());
-            oauth.setOpenid(oauthPostDto.getOpenid());
-
-            oauthRepository.save(oauth);
-            return true;
-        } else if (oauthRepository.findByOpenid(oauthPostDto.getOpenid()).isPresent()) {
-            return true;
-        }
-
-        Users user = new Users();
-
-        user.setEmail(email);
-        user.setUsername(username);
-        user.setPassword(passwordEncoder.encode(oauthPostDto.getOpenid()));
-        user.setRole(rolesRepository.findByRoleName(StaticConfig.DefaultRole).orElseThrow());
-
-        usersRepository.save(user);
-
-        Oauth oauth = new Oauth();
-
-        oauth.setUsers(user);
-        oauth.setOauthType(oauthPostDto.getOauthType());
-        oauth.setOpenid(oauthPostDto.getOpenid());
-
-        oauthRepository.save(oauth);
-        return true;
-    }
-
-    @Transactional
     public UsersGetDto updateLastLoginTimestamp(String username) {
         Users user = usersRepository.findByUsername(username).orElseThrow(() ->
                 new UsernameNotFoundException(String.format("Username %s not found", username)));
@@ -129,11 +75,6 @@ public class UsersService implements UserDetailsService {
 
     public boolean uniqueUsername(String username) {
         return usersRepository.findByUsername(username).isEmpty();
-    }
-
-    public boolean hasOpenid(UUID usersId) {
-        Users user = usersRepository.findById(usersId).get();
-        return oauthRepository.findByUsers(user).isPresent();
     }
 
     @Transactional
@@ -167,32 +108,6 @@ public class UsersService implements UserDetailsService {
     }
 
     @Transactional
-    public boolean sendVerificationEmail(UUID usersId, String newEmail) {
-        VerificationEmailMessage message = new VerificationEmailMessage(usersId, newEmail);
-        rabbitTemplate.convertAndSend("VerificationEmail", message);
-        return true;
-    }
-
-    @Transactional
-    @RabbitListener(queues = "VerificationEmail")
-    public void handleVerificationEmail(VerificationEmailMessage message) throws Exception {
-        Users user = usersRepository.findById(message.getUsersId())
-                                    .orElseThrow(() -> new UserNotFoundException("Invalid UserID"));
-
-        Util.createMail(
-                StaticConfig.fromEmail,
-                user.getEmail(),
-                "Change Email Request",
-                "Your account " + user.getUsername() + " is changing email to " + message.getNew_email());
-
-        Util.createMail(
-                StaticConfig.fromEmail,
-                message.getNew_email(),
-                "Verify Email",
-                StaticConfig.VerifyEmailContext + domainName + StaticConfig.VerifyEmailUrl + message.getNew_email());
-    }
-
-    @Transactional
     public boolean changeEmail(UUID usersId, String newEmail) {
         Users user = usersRepository.findById(usersId)
                                     .orElseThrow(() -> new UserNotFoundException("Invalid UserID"));
@@ -215,31 +130,6 @@ public class UsersService implements UserDetailsService {
         user.setPassword(passwordEncoder.encode(usersMiniPutDto.getNewPassword()));
         usersRepository.save(user);
         return true;
-    }
-
-    @Transactional
-    public boolean sendResetPasswordEmail(String email) {
-        ResetPasswordEmailMessage message = new ResetPasswordEmailMessage(email);
-        rabbitTemplate.convertAndSend("ResetPasswordEmail", message);
-        return true;
-    }
-
-    @Transactional
-    @RabbitListener(queues = "ResetPasswordEmail")
-    public void handleResetPasswordEmail(ResetPasswordEmailMessage message) throws Exception {
-        Optional<Users> user = usersRepository.findByEmail(message.getEmail());
-
-        if (user.isPresent()) {
-
-            String encode = Util.UrlEncoder(StaticConfig.JwtPrefix + Util.generateJwt(new JwtUser(user.get())));
-
-            Util.createMail(
-                    StaticConfig.fromEmail,
-                    message.getEmail(),
-                    "Reset password",
-                    StaticConfig.ResetPasswordContext +
-                            domainName + StaticConfig.ResetPasswordUrl + encode);
-        }
     }
 
     @Transactional
